Parse evdev force feedback capabilities
 * This is just the parsing part of the process, which is both the
fragile bit (the in-memory layout is different based on byte order
and machine word size) and the automatically-testable bit.

 * We go to the trouble of decoding all of the known ffbits values,
since it's easy enough to add and might come in handy later.  Also
means that nobody will be surprised by some capability being shown
as provided by the kernel but not reflected in the device in Lisp.

 * Not actually tested on a big-endian system, 32-bit or 64-bit,
but should, in theory, both pass the tests and match the data as
the kernel would provide it (including in 32-bit compatibility
mode on a 64-bit kernel).