Query evdev force feedback capabilities when opening devices
 * This is a simple wrapper around a call to IOCTL and the new
effect capability parser, plus wiring up a new slot on the device
class that's initialized by calling the wrapper.

 * I'm not completely satisfied by GET-EFFECT-CAPABILITIES but
haven't yet been able to come up with anything better.  Mainly,
the magic numbers are bothersome.

 * No test cases for this, because they would require having a
suitable gamepad connected, accessible, and not otherwise in use,
and wouldn't tell us much beyond that the IOCTL is being called
and with the appropriate IOCTL code.  Alternately, they could be
testing how things are handled when called on a non-evdev file
descriptor, which shouldn't ever be the case in the first place,
or when some other error condition (the hardware is disconnected
between opening the device node and calling ioctl(2), for example)
occurs.  Basically, this is a thin integration function, and must
be tested interactively.

 * In principle, checking that the device class has the new slot
and it is initialized correctly could be automatically tested, but
would require some changes to the existing design in order to
provide the required hooks, and wouldn't tell us much.